% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{adulting}
\alias{adulting}
\title{adulting}
\usage{
adulting(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble (main table)}

\item{y}{A data frame or tibble (rubber-stamping table)}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
A redundant function to make your code sound harder than it actually is (right_join equivalent)
}
