% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{all_that}
\alias{all_that}
\title{all_that}
\usage{
all_that(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to count by}
}
\value{
A data frame with counts
}
\description{
See if your variable is all that or not (count equivalent)
}
\examples{
mtcars |> all_that(cyl, gear)
}
