% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{cringe}
\alias{cringe}
\title{cringe}
\usage{
cringe(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A data frame with matching rows
}
\description{
Only cool kids allowed (semi_join equivalent)
}
