% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{epic_fail}
\alias{epic_fail}
\title{epic_fail}
\usage{
epic_fail(default = NULL, f, quiet = FALSE)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs (new_name = old_name)}
}
\value{
A data frame with renamed columns
}
\description{
Couldn't have been worse (failwith equivalent)
}
