% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{FOMO}
\alias{FOMO}
\title{FOMO}
\usage{
FOMO(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to group by}
}
\value{
A grouped data frame
}
\description{
Make sure you don't miss out! (group_by equivalent)
}
\examples{
mtcars |> FOMO(cyl) |> tldr(avg_mpg = mean(mpg))
}
