% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{glow_up}
\alias{glow_up}
\title{glow_up}
\usage{
glow_up(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs of expressions}
}
\value{
A data frame with new/modified columns
}
\description{
Give your columns a glow up! (mutate equivalent)
}
\examples{
mtcars |> glow_up(mpg_squared = mpg^2, high_hp = hp > 150)
}
