% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{GOAT}
\alias{GOAT}
\title{GOAT}
\usage{
GOAT(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to sort by}
}
\value{
A sorted data frame
}
\description{
See who comes out on top! (arrange equivalent)
}
\examples{
mtcars |> GOAT(desc(mpg), hp)
}
