% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{humblebrag}
\alias{humblebrag}
\title{humblebrag}
\usage{
humblebrag(x, ...)
}
\arguments{
\item{x}{An object to explain}

\item{...}{Other parameters possibly used by generic}
}
\value{
The first argument, invisibly
}
\description{
More info than we needed to be honest (explain equivalent)
}
