% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{level_up}
\alias{level_up}
\title{level_up}
\usage{
level_up(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs of expressions}
}
\value{
A data frame with new/modified columns
}
\description{
Level up your columns. (mutate equivalent)
}
\examples{
mtcars |> level_up(mpg_squared = mpg^2, high_hp = hp > 150)
}
