% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{literally}
\alias{literally}
\title{literally}
\usage{
literally(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs (new_name = old_name)}
}
\value{
A data frame with renamed columns
}
\description{
Do you mean literally or literally? (rename equivalent)
}
\examples{
mtcars |> literally(miles_per_gallon = mpg, horse_power = hp)
}
