% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{mood}
\alias{mood}
\title{mood}
\usage{
mood(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Columns to keep}
}
\value{
A data frame with selected columns
}
\description{
What columns are you in the mood for? (select equivalent)
}
\examples{
mtcars |> mood(mpg, cyl, hp)
}
