% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{quiet_quit}
\alias{quiet_quit}
\title{quiet_quit}
\usage{
quiet_quit(..., .ptype = NULL, .size = NULL)
}
\arguments{
\item{...}{One or more vectors.}

\item{.ptype}{An optional prototype declaring the desired output type.}

\item{.size}{An optional size declaring the desired output size.}
}
\value{
A data frame with matching rows
}
\description{
Let the other columns do the work (coalesce equivalent)
}
\examples{
y <- c(1, 2, NA, NA, 5); z <- c(NA, NA, 3, 4, 5); quiet_quit(y, z)
}
