% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{shook}
\alias{shook}
\title{shook}
\usage{
shook(.data)
}
\arguments{
\item{.data}{A grouped data frame}
}
\value{
An ungrouped data frame
}
\description{
Shock everyone! (ungroup equivalent)
}
\examples{
mtcars |> FOMO(cyl) |> shook()
}
