% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{sike}
\alias{sike}
\title{sike}
\usage{
sike(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A data frame with non-matching rows
}
\description{
jk (anti_join equivalent)
}
