% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{sup}
\alias{sup}
\title{sup}
\usage{
sup(.data, n = 6)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{n}{Number of rows to keep}
}
\value{
A data frame with first n rows
}
\description{
What's up? (head/slice_head equivalent)
}
\examples{
mtcars |> whats_up(5)
}
