% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{thats_sick}
\alias{thats_sick}
\title{sick}
\usage{
thats_sick(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Logical predicates to keep rows}
}
\value{
A data frame with filtered rows removed
}
\description{
Filter out the sketchy rows! (filter equivalent)
}
\examples{
mtcars |> thats_sick(mpg > 20, cyl == 4)
}
