% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{thirsty}
\alias{thirsty}
\title{thirsty}
\usage{
thirsty(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A merged data frame
}
\description{
A little desperate if you ask me (full_join equivalent)
}
