% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{THIS}
\alias{THIS}
\title{THIS}
\usage{
THIS(.data, var)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{var}{Column to extract}
}
\value{
A vector
}
\description{
Exactly this (pull equivalent)
}
\examples{
mtcars |> bootstraps(mpg)
}
