% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{tight}
\alias{tight}
\title{tight}
\usage{
tight(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
Are your dataframes tight with each other? (inner join equivalent)
}
