% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{tldr}
\alias{tldr}
\title{tldr}
\usage{
tldr(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs of summary expressions}
}
\value{
A data frame with summary statistics
}
\description{
There's too much data, just get the summary stats! (summarize equivalent)
}
\examples{
mtcars |> tldr(avg_mpg = mean(mpg), max_hp = max(hp))
}
