% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{YOLO}
\alias{YOLO}
\title{YOLO}
\usage{
YOLO(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Optional columns to determine uniqueness}
}
\value{
A data frame with unique rows
}
\description{
You only live once (distinct equivalent)
}
\examples{
mtcars |> YOLO(cyl)
}
