% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{epic_fail}
\alias{epic_fail}
\title{epic_fail}
\usage{
epic_fail(x, y)
}
\arguments{
\item{x}{Vector to modify}

\item{y}{Value or vector to compare against. When x and y are equal, the value in x will be replaced with NA.}
}
\value{
A modified version of x that replaces any values that are equal to y with NA.
}
\description{
Well that didn't work (na_if equivalent)
}
