% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennialplyr.R
\name{humblebrag}
\alias{humblebrag}
\title{humblebrag}
\usage{
humblebrag(name = "none", ...)
}
\arguments{
\item{.data}{A grouped data frame}
}
\value{
An ungrouped data frame
}
\description{
Ugh, your problem is so complex that you need everyone to know about it (create_progress_bar equivalent)
}
